requirements.txt
djangorestframework==3.12.4
Django==3.2.9
